+++ lib/src/main/java/growthbook/sdk/java/internal/services/ConditionEvaluator.java
@@ -1,4 +1,4 @@
-package growthbook.sdk.java.services;
+package growthbook.sdk.java.internal.services;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
