+++ lib/src/main/java/growthbook/sdk/java/internal/services/FeatureEvaluator.java
@@ -1,4 +1,4 @@
-package growthbook.sdk.java.services;
+package growthbook.sdk.java.internal.services;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
