+++ lib/src/main/java/growthbook/sdk/java/internal/services/IConditionEvaluator.java
@@ -1,4 +1,4 @@
-package growthbook.sdk.java.services;
+package growthbook.sdk.java.internal.services;
 
 interface IConditionEvaluator {
     Boolean evaluateCondition(String attributesJsonString, String conditionJsonString);
