+++ lib/src/main/java/growthbook/sdk/java/internal/services/MathUtils.java
@@ -1,4 +1,4 @@
-package growthbook.sdk.java.services;
+package growthbook.sdk.java.internal.services;
 
 import java.math.BigInteger;
 import java.util.List;
